HID: microsoft: map Xbox Series X buttons correctly
The Xbox Series X controller connected over bluetooth maps to several
wrong buttons/axes. For the right joystick the axes are swapped to
RX/RY. The left and right triggers are correctly remapped from gas and
break. These mappings correspond to the xpad driver when the controller
is plugged in over USB.

Signed-off-by: Jelle van der Waa <jvanderwaa@redhat.com>