platform/x86: lg: Add charge_types support
Introduce charge_types support as replacement for battery_care_limit and
alternative to the charge_control_end_threshold as applications cannot
know that only 80 or 100 is a valid. With `charge_types` exposed an
application such as UPower can see if "Long Life" is an option and then
enable it to conserve battery.

Signed-off-by: Jelle van der Waa <jvanderwaa@redhat.com>