platform/x86: Add Acer battery control driver
Some Acer laptops can configure battery related features through Acer
Care Center on Windows. This driver uses the power supply extension to
set a battery charge limit, calibration mode and exposes the battery
temperature.

This driver is based on the existing acer-wmi-battery project on GitHub
which was tested on an Acer Swift 3 (SF314-34).

Signed-off-by: Jelle van der Waa <jvanderwaa@redhat.com>